refactor(story-close): extract shared format-autofix helpers (refs #3332)
Both format-autofix forks (whole-tree and scoped) carried their own
copies of the porcelain-status parse, the formatter-command resolution,
and the stage->commit->rev-parse sequence. Hoist all three into a new
format-autofix-shared.js (listDirtyPaths, resolveFormatterCmd,
commitDirtyPaths); the two forks now differ only in file-scope, commit
subject, and log level.

Also resolve the open-ended migration-promise comment in the gates.js
autofix path: rephrase the scoped-vs-whole-tree relationship as a strict
pre-pass (never a replacement) so it no longer reads as a legacy path
awaiting removal.
